Move INFINITE_U_TURN_COSTS from TurnWeighting to Weighting
