+++ core/src/main/java/com/graphhopper/routing/weighting/Weighting.java
@@ -29,6 +29,8 @@ import com.graphhopper.util.EdgeIteratorState;
  */
 public interface Weighting {
 
+    int INFINITE_U_TURN_COSTS = -1;
+
     /**
      * Used only for the heuristic estimation in A*
      *
